Move crate::env to crate::cargo
